+++ flirextractor/errors.py
@@ -1,0 +1,8 @@
+class FlirExtractorError(Exception):
+    """Base exception for the flirextractor library"""
+    pass
+
+class UnsupportedImageError(FlirExtractorError):
+    """Thrown when the input image is not supported by this program.
+    """
+    pass