+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2020-11-24
+
+### Fixed
+
+- Improved Error thrown by `get_thermal_batch()` when a non-radiometric
+  image is passed.
+
 ## [1.0.2] - 2020-07-09
 
 ### Fixed